+++ pywincffi/kernel32/events.py
@@ -5,12 +5,12 @@ Events
 A module containing Windows functions for working with events.
 """
 
-from six import string_types, integer_types
+from six import integer_types, text_type
 
 from pywincffi.core import dist
-from pywincffi.core.checks import Enums, input_check, error_check
+from pywincffi.core.checks import Enums, input_check, error_check, NoneType
 from pywincffi.exceptions import WindowsAPIError
-from pywincffi.util import string_to_cdata
+from pywincffi.wintypes import HANDLE, SECURITY_ATTRIBUTES, wintype_to_cdata
 
 
 def CreateEvent(
@@ -30,17 +30,19 @@ def CreateEvent(
     :param bool bInitialState:
         If True the initial state will be 'signaled'.
 
-    :keyword struct lpEventAttributes:
-        A pointer to a ``SECURITY_ATTRIBUTES`` structure.  If not provided
-        then by default the handle cannot be inherited by a subprocess.
+    :keyword :class:`pywincffi.wintypes.SECURITY_ATTRIBUTES` lpEventAttributes:
+        If not provided then, by default, the handle cannot be inherited
+        by a subprocess.
 
     :keyword str lpName:
+        Type is ``unicode`` on Python 2, ``str`` on Python 3.
         The optional case-sensitive name of the event.  If not provided then
         the event will be created without an explicit name.
 
     :returns:
-        Returns a handle to the event.  If an event by the given name already
-        exists then it will be returned instead of creating a new event.
+        Returns a :class:`pywincffi.wintypes.HANDLE` to the event. If an event
+        by the given name already exists then it will be returned instead of
+        creating a new event.
     """
     input_check("bManualReset", bManualReset, bool)
     input_check("bInitialState", bInitialState, bool)
@@ -50,18 +52,15 @@ def CreateEvent(
     if lpName is None:
         lpName = ffi.NULL
     else:
-        input_check("lpName", lpName, string_types)
-        lpName = string_to_cdata(lpName)
+        input_check("lpName", lpName, text_type)
 
-    if lpEventAttributes is None:
-        lpEventAttributes = ffi.NULL
-    else:
-        input_check(
-            "lpEventAttributes", lpEventAttributes,
-            allowed_types=Enums.SECURITY_ATTRIBUTES)
+    input_check(
+        "lpEventAttributes", lpEventAttributes,
+        allowed_types=(SECURITY_ATTRIBUTES, NoneType)
+    )
 
     handle = library.CreateEvent(
-        lpEventAttributes,
+        wintype_to_cdata(lpEventAttributes),
         ffi.cast("BOOL", bManualReset),
         ffi.cast("BOOL", bInitialState),
         lpName
@@ -73,7 +72,7 @@ def CreateEvent(
         if error.errno != library.ERROR_ALREADY_EXISTS:
             raise
 
-    return handle
+    return HANDLE(handle)
 
 
 def OpenEvent(dwDesiredAccess, bInheritHandle, lpName):
@@ -89,23 +88,24 @@ def OpenEvent(dwDesiredAccess, bInheritHandle, lpName):
 
     :param bool bInheritHandle:
     :param str lpName:
+        Type is ``unicode`` on Python 2, ``str`` on Python 3.
 
     :return:
-        Returns the
+        Returns a :class:`pywincffi.wintypes.HANDLE` to the event.
     """
     input_check("dwDesiredAccess", dwDesiredAccess, integer_types)
     input_check("bInheritHandle", bInheritHandle, bool)
-    input_check("lpName", lpName, string_types)
+    input_check("lpName", lpName, text_type)
 
     ffi, library = dist.load()
 
     handle = library.OpenEvent(
         ffi.cast("DWORD", dwDesiredAccess),
         ffi.cast("BOOL", bInheritHandle),
-        string_to_cdata(lpName)
+        lpName
     )
     error_check("OpenEvent")
-    return handle
+    return HANDLE(handle)
 
 
 def ResetEvent(hEvent):
@@ -116,12 +116,12 @@ def ResetEvent(hEvent):
 
         https://msdn.microsoft.com/en-us/library/ms684305
 
-    :param handle hEvent:
+    :param :class:`pywincffi.wintypes.HANDLE` hEvent:
         A handle to the event object to be reset. The handle must
         have the ``EVENT_MODIFY_STATE`` access right.
     """
-    input_check("hEvent", hEvent, Enums.HANDLE)
+    input_check("hEvent", hEvent, HANDLE)
 
     _, library = dist.load()
-    code = library.ResetEvent(hEvent)
+    code = library.ResetEvent(wintype_to_cdata(hEvent))
     error_check("ResetEvent", code=code, expected=Enums.NON_ZERO)