+++ pywincffi/kernel32/process.py
@@ -22,6 +22,7 @@ from pywincffi.core.checks import Enums, input_check, error_check
 from pywincffi.exceptions import WindowsAPIError, PyWinCFFINotImplementedError
 from pywincffi.kernel32.handle import CloseHandle
 from pywincffi.kernel32.synchronization import WaitForSingleObject
+from pywincffi.wintypes import HANDLE, wintype_to_cdata
 
 RESERVED_PIDS = set([0, 4])
 
@@ -122,18 +123,18 @@ def GetExitCodeProcess(hProcess):
 
         https://msdn.microsoft.com/en-us/library/ms683189
 
-    :param handle hProcess:
-        The handle of the process to retrieve the exit code for
+    :param :class:`pywincffi.wintypes.HANDLE` hProcess:
+        The handle of the process to retrieve the exit code for.
 
     :returns:
         Returns the exit code of the requested process if one
         can be found.
     """
-    input_check("hProcess", hProcess, Enums.HANDLE)
+    input_check("hProcess", hProcess, HANDLE)
 
     ffi, library = dist.load()
     lpExitCode = ffi.new("LPDWORD")
-    code = library.GetExitCodeProcess(hProcess, lpExitCode)
+    code = library.GetExitCodeProcess(wintype_to_cdata(hProcess), lpExitCode)
     error_check("GetExitCodeProcess", code=code, expected=Enums.NON_ZERO)
     return lpExitCode[0]
 
@@ -156,9 +157,9 @@ def OpenProcess(dwDesiredAccess, bInheritHandle, dwProcessId):
         The id of the local process to be opened.
 
     :returns:
-        Returns a handle to the opened process in the form of
-        a void pointer.  This value can be used by other functions
-        such as :func:`TerminateProcess`
+        Returns a :class:`pywincffi.wintypes.HANDLE` to the opened process.
+        This value can be used by other functions such as
+        :func:`TerminateProcess`.
     """
     input_check("dwDesiredAccess", dwDesiredAccess, integer_types)
     input_check("bInheritHandle", bInheritHandle, bool)
@@ -171,7 +172,7 @@ def OpenProcess(dwDesiredAccess, bInheritHandle, dwProcessId):
         ffi.cast("DWORD", dwProcessId)
     )
     error_check("OpenProcess")
-    return handle
+    return HANDLE(handle)
 
 
 def GetCurrentProcess():
@@ -188,10 +189,10 @@ def GetCurrentProcess():
         produced by this function will produce an exception.
 
     :returns:
-        The handle to the current process.
+        The :class:`pywincffi.wintypes.HANDLE` to the current process.
     """
     _, library = dist.load()
-    return library.GetCurrentProcess()
+    return HANDLE(library.GetCurrentProcess())
 
 
 def GetProcessId(Process):  # pylint: disable=invalid-name
@@ -202,16 +203,16 @@ def GetProcessId(Process):  # pylint: disable=invalid-name
 
         https://msdn.microsoft.com/en-us/library/ms683215
 
-    :param handle Process:
-        The handle of the process to re
+    :param :class:`pywincffi.wintypes.HANDLE` Process:
+        The handle of the process.
 
     :return:
         Returns an integer which represents the pid of the given
         process handle.
     """
-    input_check("Process", Process, Enums.HANDLE)
+    input_check("Process", Process, HANDLE)
     _, library = dist.load()
-    pid = library.GetProcessId(Process)
+    pid = library.GetProcessId(wintype_to_cdata(Process))
     error_check("GetProcessId")
     return pid
 
@@ -224,18 +225,18 @@ def TerminateProcess(hProcess, uExitCode):
 
         https://msdn.microsoft.com/en-us/library/ms686714
 
-    :param handle hProcess:
+    :param :class:`pywincffi.wintypes.HANDLE` hProcess:
         A handle to the process to be terminated.
 
     :param int uExitCode:
         The exit code of the processes and threads as a result of calling
         this function.
     """
-    input_check("hProcess", hProcess, Enums.HANDLE)
+    input_check("hProcess", hProcess, HANDLE)
     input_check("uExitCode", uExitCode, integer_types)
     ffi, library = dist.load()
     code = library.TerminateProcess(
-        hProcess,
+        wintype_to_cdata(hProcess),
         ffi.cast("UINT", uExitCode)
     )
     error_check("TerminateProcess", code=code, expected=Enums.NON_ZERO)